chat

+++ view/myview/src/js/person.js
@@ -17,16 +17,18 @@ class person{
             let deleteItem=$(this).parents('.item');
 			let _id=$(this).data('id');
 			event.preventDefault();
-			$.ajax({
-	            type:"DELETE",
-	            url:"/person/publish?id=" + _id
-	        })
-	        .done(function (results) {
-	            if(results.success == 1){
-	            	// layer.open()
-	                deleteItem.remove();
-	            }
-	        })
+			layer.confirm('你确定要删除该条信息吗？', {
+						  btn: ['确定','取消'] //按钮
+					}, function(){
+						$.ajax({
+				            type:"DELETE",
+				            url:"/person/publish?id=" + _id
+				        })
+				        .done(function (results) {
+								layer.msg('删除成功', {time: 800});
+								deleteItem.remove();
+				        })
+					});
 		});
 	}
 }
